dedupe get_activations_by_index()
