+++ apps/inference/neuronpedia_inference/shared.py
@@ -4,7 +4,10 @@ from typing import Any
 
 import einops
 import torch
-from transformer_lens import HookedTransformer
+from transformer_lens import ActivationCache, HookedTransformer
+
+from neuronpedia_inference.config import Config
+from neuronpedia_inference.sae_manager import SAEManager
 
 request_lock = asyncio.Lock()
 
@@ -135,3 +138,20 @@ def calculate_per_source_dfa(
 
 def get_layer_num_from_sae_id(sae_id: str) -> int:
     return int(sae_id.split("-")[0]) if not sae_id.isdigit() else int(sae_id)
+
+
+def get_activations_by_index(
+    sae_type: str,
+    selected_source: str,
+    cache: ActivationCache,
+    hook_name: str,
+) -> torch.Tensor:
+    """Get activations by index for a specific layer and SAE type."""
+    if sae_type == "neurons":
+        mlp_activation_data = cache[hook_name].to(Config.get_instance().device)
+        return torch.transpose(mlp_activation_data[0], 0, 1)
+    activation_data = cache[hook_name].to(Config.get_instance().device)
+    feature_activation_data = (
+        SAEManager.get_instance().get_sae(selected_source).encode(activation_data)
+    )
+    return torch.transpose(feature_activation_data.squeeze(0), 0, 1)
